@@ -8,24 +8,26 @@
   },
   "main": "index.js",
   "scripts": {
-    "start": "nodemon --watch ./src -e ts --exec \"npm run build && npm run serve\"",
-    "build": "npm run build-ts && npm run tslint",
+    "start": "npm run build && npm run serve",
+    "build": "npm run clean:dist && npm run build-ts && npm run tslint && npm run copy:static",
     "serve": "node dist/index.js",
     "build-ts": "tsc",
-    "tslint": "tslint -c tslint.json -p tsconfig.json"
+    "tslint": "tslint -c tslint.json -p tsconfig.json",
+    "clean:dist": "node scripts/cleanDist.js",
+    "copy:static": "node scripts/copyClient.js"
   },
   "author": "Jonathon Orsi",
   "license": "MIT",
   "dependencies": {
-    "express": "^4.15.3",
-    "socket.io": "^1.7.3"
+    "express": "^4.15.4",
+    "socket.io": "^1.7.4"
   },
   "devDependencies": {
-    "@types/express": "^4.0.36",
-    "@types/node": "^8.0.13",
-    "@types/socket.io": "^1.4.29",
-    "nodemon": "^1.11.0",
-    "tslint": "^5.5.0",
-    "typescript": "^2.4.1"
+    "@types/express": "^4.0.37",
+    "@types/node": "^8.0.28",
+    "@types/socket.io": "^1.4.30",
+    "nodemon": "^1.12.1",
+    "tslint": "^5.7.0",
+    "typescript": "^2.5.2"
   }
 }
