@@ -6,9 +6,6 @@
  *
  * */
 
-  // config file
-const config = require('../reverie.json');
-
-import Reverie from './server/Reverie';
-const reverie = new Reverie(config);
-
+import { Reverie } from './server/reverie';
+const reverie = new Reverie();
+reverie.run();
