@@ -1,4 +1,4 @@
-export default class Timer {
+export class Timer {
     private startTime: Date;
     private nextTime: Date;
     constructor () {}
